requirements.txt
Django==2.2.16
django-allauth==0.42.0
flake8==3.8.3
isort==4.3.21
mysqlclient==2.0.1
oauthlib==3.1.0
pip==20.2.2
pyflakes==2.2.0
pylint== 2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
setuptools==49.6.0
sqlparse==0.3.1




